WIP: frame_superres: Add aomenc argument
Adds a "superres-mode" argument to enable superres encoding. When set it
configures the superres_mode var in the encoder's config and can be used
to set the superres numerator to a fixed or dynamic value.

There are only two modes right now: SUPERRES_FIXED and SUPERRES_DYNAMIC.
Fixed lets the superres fraction numerator be set to 8-16, each
corresponding to the same value provided to superres-mode. 0 corresponds
to superres being disabled - which is really just SUPERRES_FIXED where
the numerator is 16. 1 sets the SUPERRES_DYNAMIC mode, which is just
random for now.

Change-Id: Ibae01c77abba2599fbf51096459d09ddca458d4f