WIP: frame_superres: Add aomenc argument
Adds "superres-mode" and "superres-numerator" arguments to configure
superres scaling.

There are three modes right now: SUPERRES_NONE, SUPERRES_FIXED and
SUPERRES_DYNAMIC. Arguments 0, 1, and 2 correspond to each.

For the none mode superres will always set its scale to 1:1.

For the fixed mode superres will scale to the provided numerator where
possible. The cumulative scale change over resizing and superres is
limited to 1/2, so it may have to back off.

For the dynamic mode, the numerator is just assigned at random.

Change-Id: Ibae01c77abba2599fbf51096459d09ddca458d4f