WIP: frame_superres: Add aomenc argument
Adds "superres-mode" and "superres-numerator" arguments to configure
superres scaling.

There are only two modes right now: SUPERRES_FIXED and SUPERRES_DYNAMIC.
0 sets the fixed mode and 1 sets the dynamic mode.

For the fixed mode superres will scale to the provided numerator where
possible. The cumulative scale change over resizing and superres is
limited to 1/2, so it may have to back off.

For the dynamic mode, the provided numerator will be used as the initial
numerator. Asides from that it's just random for now.

Change-Id: Ibae01c77abba2599fbf51096459d09ddca458d4f