Update patch set 14
Patch Set 14:

> I have some x86 and windows bugs to file for the existing set. Let
 > me get that together before we land any more of these

Those were fixed, I believe, and I've now rebased this patch and updated the patch in line with the fixes that were made (some type nitpicking by the sanitisers).  This patch is very old and it would be nice to land it now for completeness of the tests.

Patch-set: 14