A temporary fix on the RAM issue for upsampled referencing
When ext-refs is on, the encoding of HD videos, e.g. 1080p sequences
will cause the failure to allocate up-sampled frame buffers. This CL
temporarily turn off the use of upsampled references when ext-refs is
on for any video resolution larger than 720 in either dimension. A
more effective scheme for up-sampled referencing should be later
considered.

With the fix, the AWCY performance gains of ext-refs, vs baseline are:

PSNR Y
Average -2.55%
1080p   -2.81%
360p    -2.38%
720p    -3.10%

Change-Id: Ica5e75d82190cd68e64c536503e3559f1cfdbdd1