Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> so generic needs to be explicit with this change?
> 
> https://build.aomedia.org/jenkins/job/libaom__compile-cmake/43/

It's supposed to be, but I forgot to include generic in the list of CPU names I added to aom_configure.cmake, so this thing is completely broken right now. I'll get to it on Monday. Thanks for running the builds!

Patch-set: 2
Reviewer: Gerrit User 5215 <5215@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>