Add SSSE3 warp filter
The SSSE3 filter is very similar to the SSE2 filter, but
the horizontal pass is sped up by using the 8x8->16
multiplies added in SSSE3.

On the machine this patch was tested on, the timings
per 8x8 block are:
before: 320ns
after:  273ns
for an overall filter speedup of ~17%

Filter output is unchanged.

Change-Id: Ifb428a33b106d900cde1b080794796c0754ae182