Upgrade warp filter from SSE2 to SSSE3
Speed up the horizontal filter stage by using the 8x8->16
multiplies added in SSSE3.

The output of the warp filter is very slightly changed in
extreme cases, to make this possible. This should have very
little effect on BDRATE, but note that (for the moment)
this means that the lowbd path and the highbd path give
different results for 8-bit inputs.

On the machine this patch was tested on, the timings
per 8x8 block are:
before: 307ns
after:  254ns
for an overall filter speedup of ~21%

Change-Id: Ifb428a33b106d900cde1b080794796c0754ae182