Update patch set 3
Patch Set 3:

(4 comments)

I guess this is fine as a first step towards finding and transmitting an alpha.  Looking at the Cfl presentation I wonder if it would help, in particular for inter, if linear regression as in Thor is used to predict alpha (and only the difference from the alpha prediction will be signalled).  That would add slight decoder complexity, though.  But we can discuss that later when you have a patch for computing and signalling alpha.

Patch-set: 3
Reviewer: Gerrit User 5111 <5111@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>