Improve context modelling for ext-inter + ref-mv
For this combination of experiments, inter compound modes
were previously signalled using a simple tree structure with
a handful of contexts, just as for single inter modes without
ref-mv.

This patch updates the encoding to use a system which is more
like the encoding of single inter modes with ref-mv, including
support for the ALL_ZERO flag of ref-mv.

Change-Id: Ia3b173e2046360de6a8f4c11c7b393118e2cd8c4