Add encoder/decoder pipeline to support single ref comp modes
Now the single ref comp mode should work with WEDGE and
COMPOUND_SEGMENT. For motion_var, the OBMC_CAUSAL mode uses the 2nd
predictor if the neighboring block is single ref comp mode predicted.

This patch removes the use of SR_NEAR_NEWMV, as it requests the
signaling of two DRL indexes.

Change-Id: If6140455771f0f1a3b947766eccf82f23cc6b67a