Revert "warp_affine_c: Refactor highbd and lowbd versions."
This reverts commit 8cd0e7efac95844556d98d4634755b47b203a3ed.

Reason for revert:
This change breaks av1_warp_affine_c when CONFIG_HIGHBITDEPTH is enabled.

In particular, running ./test_libaom --gtest_filter=*Warp* compiled with --enable-warped-motion --enable-highbitdepth shows several test failures, followed by a segmentation fault when it gets up to test SSE2/AV1WarpFilterTest.CheckOutput/4

The tricky part is that the use the lowbd version of the function is dependent on a mix of two conditions:
(1) Compile time check for CONFIG_HIGHBITDEPTH and
(2) Run time check to see if bit-depth == 8
So, it is tricky to refactor.

Change-Id: I610c537fb65bde4f357185a13081639f906351de