[CfL] Store luma as prediction for chroma
Stores the reconstructed luma pixels for each transform block inside a
prediction block. Rectangular transform blocks are supported.

As for RDO, after all the modes have been tested for luma, an extra
encoding is perform in order to store the reconstructed pixel values of
the best mode. These values are then used for RDO on the chromatic
planes.

Change-Id: I354d9827e32fd41065f1b2ce02832d943a6fa156