Intra prediction: Remove unused variants.
Directional predictors for 45, 63 and 207 angle had 2 or 3 variants
each, and only one of them was actually being used. So, removed the
C and assembly versions of the unused ones.

Also, test_intra_pred_speed was testing the unused versions, so changed
it to use the version actually used by code. This meant updating one
golden MD5 value.

BUG=aomedia:442

Change-Id: I09725d593408b81e0cd636e70a88c28eea5f2222