resize-refactor: Refactor 2-pass VBR resizing
Replaces resizing's table and index approach to frame sizes with
numerator and denominator integers in the AV1_COMP struct.

This approach is more flexible and will be simpler going forward as it
is much more similar to the 1-pass CBR approach that is also being
refactored. The intention is to merge both approaches and this is the
first step toward that.

Change-Id: I5733c0687390f8a8e2790dcddfa09fb08ab88376