Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+1
> 
> Sorry I missed the discussion thread when doing the review.
> 
> The assertion in the bug report was fixed in
> https://aomedia-review.googlesource.com/c/10569/
> 
> It was caused by out of boundary writing several steps earlier to that assertion.

Thanks for the update. If you file bugs or otherwise address a nightly failure it's good to respond to the mail thread too, so others know what is going on. In practice for something like this it's best to revert unless the fix is singular and simple. Keeping the time the tree is broken down helps as others are merging in changes.

Patch-set: 3
Reviewer: Gerrit User 5020 <5020@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>