Update patch set 1
Patch Set 1:

I actually think this may actually make things somewhat more confusing.
A better option is to throw an error and exit for cases where a sub-experiment is on but the parent is off.
Some of the other cases - such as var-tx with rect-tx is unnecessary since the argument there is coding efficiency and not that one depends on the other.

Patch-set: 1
Reviewer: Gerrit User 5189 <5189@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>