Fix encode/decode mismatch with global/warped motion
When predicting a 4x4 warp block (either using ZEROMV with
global-motion, or the WARPED_CAUSAL motion mode with
warped-motion), the warp filter would previously write
4 bytes to the right of the block.

This caused encode/decode mismatches when encoding with
multiple threads and tile_cols > 1, since in that case
we could end up overwriting already-generated pixels from
the next tile across.

This patch changes the filter so that we only overwrite the
intended pixels.

Change-Id: I3664b44e872e85aa5ccc0a5781f0f9ad994a5b80