Update patch set 5
Patch Set 5:

Hi Sean,
Given the slowdown - I was wondering if the clamping guarantee is really necessary for hardware. We know based on the normative suitability test for the shear params that we will never hit > 192. Storing one extra row I am pretty sure is not that expensive, and you can always clamp non-normatively in your hardware implementation if you still want the guarantee.

Patch-set: 5