Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Hi Sean,
> Given the slowdown - I was wondering if the clamping guarantee is really necessary for hardware. We know based on the normative suitability test for the shear params that we will never hit > 192. Storing one extra row I am pretty sure is not that expensive, and you can always clamp non-normatively in your hardware implementation if you still want the guarantee.

The reason for this change was not the extra storage, but the extra address bit to the table. I suggest that we remove the clamp operation by filling in the last row in the table with a row identical to the one before it in each branch of the #if WARPEDPIXEL_PREC_BITS == 6. 

Shall I prepare a new CL for that?

Patch-set: 5