Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+2
> 
> Not every developers are familiar with every experiments.
> Describe the dependency in configure is a good way for them to reference to.
> 
> About feature dependency, if we don't allow that, then each experiments will have to deal with all scenarios that other experiments are on and off. As the experiments flag number grows, the number of scenarios grow exponentially.

I am not saying that we need to support all possible combinations - only the ones that make sense. The specific disallowed scenarios can be checked and we can exit from configure with an informative message if an illegal/unsupported combination is used. From the developer of the tools it takes almost the same effort either to add the soft enable in configure or to add the check in configure.

Soft enabling tools silently on the other hand can conflate testing and mix up tools inadvertently. If tool B depends on A, I would rather be forced to use --enable-B --enable-A rather than just --enable-B (that silently adds --enable-A), even though the latter is more concise.

Patch-set: 2
Reviewer: Gerrit User 5189 <5189@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>