Add a new coding tool of ext_comp_refs
This is to revisit the use of compound reference prediction:
(1) Re-design the entropy coding of the reference frames;
(2) Allow the pair of compound reference frames share the same sign
    bias;
(3) Add more compound reference options.

This is a try to leverage the existing syntax in AV1+LOWDEALY_COMPOUND
to further improve the compound ref performance.

Change-Id: I9d96ff2ad16a06af5f84547780388154e2b2d3c2