Update patch set 1
Patch Set 1:

If I follow the steps you posted in the bug report for the commit directly before this, that test still fails. The configuration in the bug report also doesn't really point to this patch because this only makes significant changes to the code base if compound_segment is enabled with ext_inter. There are a few changes to ext_inter alone so I guess it is possible this it caused a failure but they are only name changes.

Patch-set: 1
Reviewer: Gerrit User 5186 <5186@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>