cb4x4: Don't assume TX units are half MI_SIZE in predict_square_intra_block()
Currently the "transform units" used to measure transform sizes
in some parts of the code are based on the smallest defined
transform size.

cb4x4 currently defines a 2x2 transform size, even when chroma_2x2
is not enabled, which means that the scale of the transform units
was always double that of MODEINFO units.

Several areas of the code were hard-coding this assumption instead
of converting from one to the other using appropriate constants.

Change-Id: I86057b85b7d4cd810d87da28f5eedc5b27597e29