Update patch set 7
Patch Set 7:

> the issue is pre-existing and resolved by the tempmv_signaling
 > experiment. Later on, when global motion experiment is introduced,
 > it re-use the previous frame global motion parameters instead of
 > the parameters from LAST FRAME. That's why I want you to take a
 > look if there is any impact on global motion.

I can run some experiments to see how the global motion results change when we have cm->use_prev_frame_mvs == 1. I don't think it will be a big deal since the gm params for LAST_FRAME should still be similar to the parameters we are trying to code. If it makes much of a difference, I can adjust it as a followup but it seems that this patch fixes a mismatch and should be checked in.

Patch-set: 7