Remove the incorrect comment
The element of "idx" in the data structure of RefBuffer is a
connection between AV1_COMMON::RefBuffer and
AV1_COMMON::BufferPool::RefCntBuffer.

It is being used currently in many places in both the encoder and the
decoder. "idx" is not only used to track whether a reference buffer is
alive, but also serve as the connection to RefCntBuffer which contains
info that RefBuffer itself cannot access.

Change-Id: I5ccdb9c2afe840d04a025da359a704f5fc216ddd