Change confusing naming "sb" to "cb".
Many places use the naming "super_block" or "sb", when they are actually
referring to a "coding block" or "cb". This has caused confusions for
many ppl reading the code for the 1st time.

Changed this naming to 'coding block' or 'cb', mainly in rdopt.c and a
few other places.

Change-Id: I609ea0e3e838ed6eceeab187be7627687e18dcac