cb4x4: Elide special cases cb4x4 doesn't need
These conditions in has_top_right() are never true when cb4x4 is
enabled.

The corresponding check in has_bottom_left() is already behind a
similar #if.

This is needed by a subsequent patch to move sub-4x4 block sizes
behind their own configuration flag, since that changes the value
of bw_unit, which might cause the condition to become true when it
should not be.

Encoder and decoder output should remain unchanged.

Change-Id: I648654dfc265f911d83668d94dc927ebc0c70771