superres: Send scale every frame
The superres scale was accidentally only sent if the current frame was
a key frame or if the current resolution was not used by any of the
reference frames. This patch separates the superres scale from the
current resolution to simplify its coding and makes sure that it is
written for every frame.

Change-Id: I419912c833cdc9670fc5a8ea1d67393d72ce402d