Refactor hbd txfm configurations to be 1D
The hbd transform configurations were originally written for all possible
2d transforms. Now that there are many more possible 2d transforms
due to EXT_TX and RECT_TX, it is simpler to write the cfg for the
4 1D transform types and compose them to make all new possible transform
types. This will allow for an easier integration of the identity transform
for EXT_TX and rectangular transforms for RECT_TX into the current
hbd transform codepath and facilitate the removal of obsolete transforms.
This has no impact on performance.

Note: So far I have only tested performance on a few local videos.
I'd like to wait for the borg jobs to finish before submitting to
ensure that this patch doesn't impact performance for at least the
lowres set.

BUG=aomedia:524

Change-Id: I8d974e22621658d69cbbf36e330d5415562db243