Add -DPIC to ASFLAGS
The configure scripts expects the default settings of the assembler
to work with the default settings of the linker.
This isn't true on Ubuntu & Debian, where the GNU linker, by default,
is expecting position independent code (for security considerations).

This commit allows the default build to succeed on these platforms.

BUG=aomedia:102

Change-Id: I8ccca9181c3a826a6b4dc700558973fea41867f2