Update patch set 9
Patch Set 9:

A question for Sarah/Yue/Deb:

This patch makes ~1000 lines of existing code be dead code, so the net impact of https://aomedia-review.googlesource.com/c/11987/ + this + dead code removal is about -200 lines. There's also a couple of thousand lines of SSE code which will need to be replaced by a new (likely to be shorter) version.

Would it be best to merge the dead code removal into this patch, or would it be better to have that as a separate patch if/when this is merged?

Patch-set: 9
Reviewer: Gerrit User 5975 <5975@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>