frame_superres: Fix superres scale encoding
This commit makes several changes that to improve/fix the encoding of
the superres state.

The superres upscaled size is added to the common struct for the decoder
to use to store the recieved "full" frame size before replacing it with
the downsampled superres size so various buffers get reserved correctly.
It is also used by the encoder to store the resolution it will be
upscaling to post-encode to reduce interaction with resizing later.

This also resolves an issue where the superres scaling factor numerator
was not sent if the frame size was encoded by reference to a previous
frame.

These changes also lead to changing how resizing works again. Resizing
was storing the current scaled resolution in the AV1EncoderConfig
struct, which shouldn't be changed at run time so all but the initial
writes to it were removed. The current scale was not added to the
AV1_COMP struct since it's really only needed when setting the
resolution to scale to - it's used and then thrown away.

These changes make setting the downscaled size simpler and encoding the
superres state simpler as well. These changes can be seen in
setup_frame_size.

Change-Id: Iab00b2e87b9c26ee90689d9f4dc2a6cdf06d1f73