Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I'd like to understand what av1_zero_above_context() is actually intended to do. As far as I can tell, it seems like it is only ever called with mi_col_start and mi_col_end set to tile boundaries. That means that the only case where it should fail to be a multiple of mib_size_log2 is on the right edge, when the video width is not a multiple of the superblock size.
> 
> When above_context is allocated in alloccommon.c, their size is also rounded up, but there is a comment that says these arrays are bigger than they need to be. What code now actually relies on the extra entries being zero?

As far as I know, the only code which needs the extra entries to be zeroed is get_entropy_context(). This is because it always reads an amount of data corresponding to the transform width, even if the transform block runs off the edge of the frame.

For a longer explanation, see the commit message for https://aomedia-review.googlesource.com/c/11081/

Patch-set: 1
CC: Gerrit User 5975 <5975@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>