Update patch set 3
Patch Set 3:

Yes, sorry, I did not mean to imply we shouldn't make this change (so I'm glad to see it was submitted).

But this function is making some non-obvious assumptions about how it gets called (e.g., the aligned_width calculation only works because mi_col_start is itself already superblock-aligned, and only mi_col_end is allowed not to be), those assumptions are not documented, and there are not even any asserts to verify they're true. The mistake was introduced because I mistook "mib_size_log2" to be "the size of a MI block in TX units" or something similar rather than "the maximum size of a superblock in MI units", because it never would have occurred to me to think of the latter based on the function name or parameter names or how the values were being used, nor even to think that the latter wouldn't be a constant.

The first step to making that better (in follow-on patches) is to understand what is actually intended.

 > As far as I know, the only code which needs the extra entries to be
 > zeroed is get_entropy_context(). This is because it always reads an
 > amount of data corresponding to the transform width, even if the
 > transform block runs off the edge of the frame.

So, help me out here. My understanding was that we forcibly split down blocks on the bottom/right edge that exceed the size of the visible MI grid. How does one get a transform that spans beyond this edge?

Patch-set: 3
Reviewer: Gerrit User 5075 <5075@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>