DO NOT SUBMIT: Restrict Wiener filter range
This patch is to test whether we can restrict the Wiener filters
produced at the encoder side to make sure that they can't lead
to integer overflow when using the high-precision filters.

This tests using a slightly conservative but easy to check criterion,
and is designed to check if we get any loss in the "best case".

Change-Id: I58e5a215bc99b01dcd47afb2750e878b40c289cc