update the quantizer checking to enable real constant QP testing
since the current encoder always enable QP modulation even the
bit rate control is set to constant QP mode. in order to enable
real constant QP encoding, have to specify the min-q and max-q
to be the same as the target qp. but original quantizer checking
function treat (min-q==max-q) as waring and cause encoder to bail.

Change-Id: Ifceb14a0fc7c011c8fbe7f4941e9b641640d735a