Update patch set 1
Patch Set 1:

The codec will indeed function with minq and maxq the same, however it functions differently than H.264 and HEVC in this regard. This difference has caused people to publish papers where they compare VP9 or AV1 with no keyframe boost with HM with a large keyframe boost, with predictable results.

The warning as it exists in the codebase has been entirely ineffective for preventing the error, so I'd be happy to see it changed. However, instead of this change, I'd prefer an explicit warning for minq=maxq, that mentions this difference, and warns that it's almost certainly not what the user wants.

Patch-set: 1
Reviewer: Gerrit User 5050 <5050@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>