Update patch set 9
Patch Set 9:

(9 comments)

nothing here is a blocker, just some cosmetic comments that can be done in a follow up. for a change like this it would be best to land it and move forward to avoid having to deal with the potential of new calls to the function being added.

Patch-set: 9
CC: Gerrit User 5020 <5020@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>