Fix Windows x86 build with --enable-ext-inter
The Windows calling convention pushes any __m128i type arguments
after the 3rd (4th on x86-64) onto the stack. But on x86,
stack-allocated arguments are not guaranteed to be aligned to
a multiple of their natural alignment, leading to compile errors.

We fix this by making the functions which take >3 __m128i arguments
instead take pointers. Since the functions are marked INLINE, the
extra memory operations should optimize out.

BUG=aomedia:587

Change-Id: I0cb2831fd12aded6f2821c037365386e6183ba5c