Fix Windows x86 build with --enable-ext-inter
The Windows calling convention pushes any __m128i type arguments
after the 3rd (4th on x86-64) onto the stack. But on x86, the
stack is not guaranteed to be 16-byte aligned, so Visual Studio
was erroring out.

We fix this by making the functions which take >3 __m128i arguments
instead take pointers for the later arguments. Since the functions
are marked INLINE, the extra memory operations should optimize out.

Change-Id: I0cb2831fd12aded6f2821c037365386e6183ba5c