Update patch set 8
Patch Set 8:

(1 comment)

This now compiles, but with warnings which shouldn't be ignored (see comment in inv_txfm.h).

The clamp_buf function does change the decoder behaviour for existing streams, but as the change should be symmetric in the encoder I have no reason to oppose it on those grounds.

I can confirm patch set 8 still acts as a fix for:
  https://bugs.chromium.org/p/aomedia/issues/detail?id=591

Patch-set: 8